requirements.txt
absl-py==0.4.0
astor==0.7.1
gast==0.2.0
grpcio==1.14.1
Markdown==2.6.11
numpy==1.14.5
Pillow==5.2.0
protobuf==3.6.1
scipy==1.1.0
six==1.11.0
tensorflow-gpu==1.7.1
tensorflow-tensorboard==0.1.8
termcolor==1.1.0
torch==0.4.1
torchvision==0.2.1
Werkzeug==0.14.1
